api/requirements.txt
# ESG RAG API 依賴
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2

# 向量搜尋（FAISS CPU 版本，不需 GPU）
faiss-cpu>=1.9.0
numpy>=1.26.4,<2

# GCP SDK（Embedding 用 Vertex AI，Generation 用 AI Studio）
google-cloud-aiplatform>=1.60.0
google-cloud-storage>=2.17.0
vertexai>=1.60.0
google-genai>=1.0.0
google-auth>=2.23.0

# Rate limiting / caching
slowapi>=0.1.9
cachetools>=5.3.0

# HTTP 工具
httpx==0.27.0

# LINE Bot（選填，未安裝時 /webhook 端點不啟用）
line-bot-sdk>=3.11.0
